backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
celery
redis
djangorestframework-simplejwt
requests
dj-database-url
# (doublon psycopg2-binary supprimé)
